+++ src/ParsedFile.js
@@ -4,6 +4,8 @@ import esprima from 'esprima';
 import estraverse from 'estraverse';
 import TagDict from './TagDict';
 import {relative, resolve} from 'path';
+import * as ts from 'typescript';
+import {parse as parseTS} from './TypeScriptParser';
 
 export default class ParsedFile {
   /**
@@ -19,7 +21,6 @@ export default class ParsedFile {
     }
 
     this._code = params.code || '';
-    this._ast = params.ast;
     this._path = params.path;
     this._docComments = params.docComments || [];
     this._relativePath = params.relativePath;
@@ -31,10 +32,6 @@ export default class ParsedFile {
 
   get path() {
     return this._path;
-  }
-
-  get ast() {
-    return this._ast;
   }
 
   get docComments() {
@@ -61,14 +58,23 @@ export default class ParsedFile {
       });
   }
 
+  static _buildTSDocComments(comments) {
+    return comments
+      .map(comment => {
+        const text = comment.text.substring(3, comment.text.length - 2).replace(/^ *\*+/mg, '');
+        return {
+          value: text,
+          location: {start: comment.start, end: comment.end}
+        };
+      })
+      .map(comment => {
+        comment.tagdict = TagDict.parse(comment);
 
-  /**
-   * @param {string} code
-   * @param {Object} options
-   * @param {string} options.path
-   * @param {string} options.basePath
-   */
-  static parse(code, options = {}) {
+        return comment;
+      });
+  }
+
+  static _parseJS(code, options = {}) {
     const ast = esprima.parse(code, {
       loc: true,
       range: true,
@@ -83,7 +89,6 @@ export default class ParsedFile {
     const params = {
       code: code,
       path: options.path ? resolve(options.path) : undefined,
-      ast: ast,
       docComments: docComments
     };
 
@@ -92,6 +97,32 @@ export default class ParsedFile {
     }
 
     return new ParsedFile(params);
+  }
+
+  /**
+   * @param {string} code
+   * @param {Object} options
+   * @param {string} options.path
+   * @param {string} options.basePath
+   */
+  static parse(code, options = {}) {
+    if (typeof code !== 'string') {
+      throw Error('"code" parameter should be a string.');
+    }
+
+    if (options.path && options.path.endsWith('.ts')) {
+      const relativePath = relative(resolve(options.basePath), options.path);
+      const comments = ParsedFile._buildTSDocComments(parseTS(code));
+
+      return new ParsedFile({
+        path: options.path,
+        relativePath: relativePath,
+        docComments: comments,
+        code: code
+      });
+    } else {
+      return ParsedFile._parseJS(code, options);
+    }
   }
 }